requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==24.2
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
django-storages==1.14.2
djangorestframework==3.15.1
google==3.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-oauth==1.0.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.6
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
